rename scheduled workflow modal folder and file
